Hmmm

+++ src/days/day19.rs
@@ -22,9 +22,9 @@ struct State {
     time: u8,
     geodes: u8,
 
-    ore: u16,
-    clay: u16,
-    obsidian: u16,
+    ore: u8,
+    clay: u8,
+    obsidian: u8,
 
     ore_robot: u8,
     clay_robot: u8,
@@ -59,16 +59,16 @@ fn solve(blueprint: &Blueprint, end_time: i32, mut max_time: Option<&mut i32>) -
             return candidates;
         }
 
-        if s.ore >= blueprint.geode_robot_ore as u16
-            && s.obsidian >= blueprint.geode_robot_obsidian as u16
+        if s.ore >= blueprint.geode_robot_ore as u8
+            && s.obsidian >= blueprint.geode_robot_obsidian as u8
         {
             candidates.push((
                 State {
                     time: s.time + 1,
-                    ore: s.ore + s.ore_robot as u16 - blueprint.geode_robot_ore as u16,
-                    clay: s.clay + s.clay_robot as u16,
-                    obsidian: s.obsidian + s.obsidian_robot as u16
-                        - blueprint.geode_robot_obsidian as u16,
+                    ore: s.ore + s.ore_robot as u8 - blueprint.geode_robot_ore as u8,
+                    clay: s.clay + s.clay_robot as u8,
+                    obsidian: s.obsidian + s.obsidian_robot as u8
+                        - blueprint.geode_robot_obsidian as u8,
                     geodes: s.geodes + s.geode_robots,
                     ore_robot: s.ore_robot,
                     clay_robot: s.clay_robot,
@@ -79,15 +79,15 @@ fn solve(blueprint: &Blueprint, end_time: i32, mut max_time: Option<&mut i32>) -
             ));
         }
 
-        if s.ore >= blueprint.obsidian_robot_ore as u16
-            && s.clay >= blueprint.obsidian_robot_clay as u16
+        if s.ore >= blueprint.obsidian_robot_ore as u8
+            && s.clay >= blueprint.obsidian_robot_clay as u8
         {
             candidates.push((
                 State {
                     time: s.time + 1,
-                    ore: s.ore + s.ore_robot as u16 - blueprint.obsidian_robot_ore as u16,
-                    clay: s.clay + s.clay_robot as u16 - blueprint.obsidian_robot_clay as u16,
-                    obsidian: s.obsidian + s.obsidian_robot as u16,
+                    ore: s.ore + s.ore_robot as u8 - blueprint.obsidian_robot_ore as u8,
+                    clay: s.clay + s.clay_robot as u8 - blueprint.obsidian_robot_clay as u8,
+                    obsidian: s.obsidian + s.obsidian_robot as u8,
                     geodes: s.geodes + s.geode_robots,
                     ore_robot: s.ore_robot,
                     clay_robot: s.clay_robot,
@@ -98,13 +98,13 @@ fn solve(blueprint: &Blueprint, end_time: i32, mut max_time: Option<&mut i32>) -
             ));
         }
 
-        if s.ore >= blueprint.clay_robot_ore as u16 {
+        if s.ore >= blueprint.clay_robot_ore as u8 {
             candidates.push((
                 State {
                     time: s.time + 1,
-                    ore: s.ore + s.ore_robot as u16 - blueprint.clay_robot_ore as u16,
-                    clay: s.clay + s.clay_robot as u16,
-                    obsidian: s.obsidian + s.obsidian_robot as u16,
+                    ore: s.ore + s.ore_robot as u8 - blueprint.clay_robot_ore as u8,
+                    clay: s.clay + s.clay_robot as u8,
+                    obsidian: s.obsidian + s.obsidian_robot as u8,
                     geodes: s.geodes + s.geode_robots,
                     ore_robot: s.ore_robot,
                     clay_robot: s.clay_robot + 1,
@@ -115,13 +115,13 @@ fn solve(blueprint: &Blueprint, end_time: i32, mut max_time: Option<&mut i32>) -
             ));
         }
 
-        if s.ore >= blueprint.ore_robot_ore as u16 {
+        if s.ore >= blueprint.ore_robot_ore as u8 {
             candidates.push((
                 State {
                     time: s.time + 1,
-                    ore: s.ore + s.ore_robot as u16 - blueprint.ore_robot_ore as u16,
-                    clay: s.clay + s.clay_robot as u16,
-                    obsidian: s.obsidian + s.obsidian_robot as u16,
+                    ore: s.ore + s.ore_robot as u8 - blueprint.ore_robot_ore as u8,
+                    clay: s.clay + s.clay_robot as u8,
+                    obsidian: s.obsidian + s.obsidian_robot as u8,
                     geodes: s.geodes + s.geode_robots,
                     ore_robot: s.ore_robot + 1,
                     clay_robot: s.clay_robot,
@@ -135,9 +135,9 @@ fn solve(blueprint: &Blueprint, end_time: i32, mut max_time: Option<&mut i32>) -
         candidates.push((
             State {
                 time: s.time + 1,
-                ore: s.ore + s.ore_robot as u16,
-                clay: s.clay + s.clay_robot as u16,
-                obsidian: s.obsidian + s.obsidian_robot as u16,
+                ore: s.ore + s.ore_robot as u8,
+                clay: s.clay + s.clay_robot as u8,
+                obsidian: s.obsidian + s.obsidian_robot as u8,
                 geodes: s.geodes + s.geode_robots,
                 ore_robot: s.ore_robot,
                 clay_robot: s.clay_robot,
